what changes have been made

+++ C_GroupProj/encrypt.h
@@ -6,4 +6,4 @@ int xor(int);
 
 char asctochar(int);
 
-unsigned long encrypt(const char*, const unsigned int, const unsigned int);
+long long encrypt(const char*, const unsigned int, const unsigned int);
